Move determine_content_type from gcs_storage to storage_common utils
This'll be used in more places than GCS storage, so make it generic!
